requirements.txt
fastapi 
uvicorn 
openai 
pytesseract 
pillow 
requests 
